+++ lib/src/time_ago_provider.dart
@@ -1,6 +1,12 @@
+import 'package:time_ago_provider/time_ago_provider.dart';
+
 import 'languages/language.dart';
 import 'languages/english.dart';
 import 'languages/german.dart';
+import 'languages/polish.dart';
+import 'languages/arabic.dart';
+import 'languages/french.dart';
+import 'languages/turkish.dart';
 
 /// Formats provided [date] to a fuzzy time like '9 minutes ago'
 /// - If [locale] is passed will look for message for that locale, if you want
@@ -72,7 +78,17 @@ Map<String, Language> _languages = {
   'en': English(),
   'en_short': English(shortForm: true),
   'de': German(),
-  'de_short': German(shortForm: true)
+  'de_short': German(shortForm: true),
+  'pl': Polish(),
+  'pl_short': Polish(shortForm: true),
+  'fr': French(),
+  'fr_short': French(shortForm: true),
+  'ar': Arabic(),
+  'ar_short': Arabic(shortForm: true),
+  'tr': Turkish(),
+  'tr_short': Turkish(shortForm: true),
+  'pr': Portuguese(),
+  'pr_short': Portuguese(shortForm: true)
 };
 
 /// Sets a [locale] with the provided [Language] to be available when