+++ lib/time_ago_provider.dart
@@ -2,7 +2,10 @@ library time_ago_provider;
 
 export 'src/languages/english.dart';
 export 'src/languages/german.dart';
+export 'src/languages/polish.dart';
 export 'src/languages/french.dart';
 export 'src/languages/arabic.dart';
+export 'src/languages/turkish.dart';
+export 'src/languages/portuguese.dart';
 
 export 'src/time_ago_provider.dart';